backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.19
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
insightface==0.7.3
onnxruntime==1.19.2
faiss-cpu==1.9.0
torch==2.5.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
